+++ CHANGELOG.md
@@ -75,6 +75,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
   - Added a new task, `update-cmr-access-constraints`, that will set access constraints in CMR Metadata.
     Currently supports UMMG-JSON and Echo10XML, where it will configure `AccessConstraints` and
     `RestrictionFlag/RestrictionComment`, respectively.
+- **CUMULUS-2112**
+  - Added `@cumulus/api/lambdas/internal-reconciliation-report`, so create-reconciliation-report
+    lambda can create `Internal` reconciliation report
 - **CUMULUS-2116**
   - Added `@cumulus/api/models/granule.unpublishAndDeleteGranule` which unpublishes a granule from CMR and deletes it from Cumulus, but does not update the record to `published: false` before deletion
 - **CUMULUS-2113**